move non-shared functions from vmi_util to builder classes
