+++ lukefi/metsi/data/formats/nfi/vmi9_builder.py
@@ -110,7 +110,7 @@ class VMI9Builder(VMIBuilder):
         result.municipality_id = util.parse_int(vmi_util.vmi_codevalue(row["municipality"]))
         result.auxiliary_stand = row["stand_number"] != '1'
 
-        parsed = vmi_util.parse_vmi12_date(row["date"])
+        parsed = vmi_util.parse_date(row["date"])
         if parsed is None:
             raise MetsiException("Year is None in VMI9 data")
 
@@ -398,16 +398,16 @@ class VMI9Builder(VMIBuilder):
         identifier = f"{stand_identifier}-{i + 1}-stratum"
 
         strata.identifier[i] = identifier
-        strata.species[i] = int(species)
+        strata.species[i] = species
         strata.mean_diameter[i] = mean_diameter
         strata.mean_height[i] = mean_height
         strata.breast_height_age[i] = vmi_util.parse_int0(seg_d13_age_raw)
         strata.biological_age[i] = age
         strata.stems_per_ha[i] = stems_per_ha
         strata.basal_area[i] = basal_area
-        strata.origin[i] = int(syntytapa)
+        strata.origin[i] = syntytapa
         strata.stratum_number[i] = i + 1
-        strata.storey[i] = int(storey)
+        strata.storey[i] = storey
         strata.sapling_stems_per_ha[i] = 0.0
         strata.number_of_generated_trees[i] = 0
 
@@ -495,7 +495,7 @@ class VMI9Builder(VMIBuilder):
 
         trees.identifier[i] = identifier
         trees.tree_number[i] = tree_number
-        trees.species[i] = int(species)
+        trees.species[i] = species
         if tree_category is not None:
             trees.tree_category[i] = tree_category
         trees.breast_height_diameter[i] = breast_height_diameter
